requirements.txt
# Core Flask and web framework
Flask==2.3.3
Werkzeug==2.3.7

# Computer Vision and Face Recognition
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.0.1

# Database
mysql-connector-python==8.1.0

# Security
bcrypt==4.0.1

# Production server
waitress==2.1.2

# Development tools (optional)
python-dotenv==1.0.0
